requirements.txt
streamlit == 1.51.*
google-adk == 1.18.*
